requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
hmmlearn>=0.3.0
statsmodels>=0.14.0

# Financial data and options
yfinance>=0.2.28
QuantLib-Python>=1.31

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Network analysis
networkx>=3.1
graph-tool>=2.45; sys_platform != 'darwin'  # Skip on macOS due to build complexity

# Data visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web and API
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
supabase>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
